% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_humidity.R
\name{absolute_humidity}
\alias{absolute_humidity}
\title{Calculate the absolute humidity in 1 m^3 dependence of the temperature and the relative humidity}
\usage{
absolute_humidity(temperature, relative_humidity = 100)
}
\arguments{
\item{temperature}{temperate in degress celsius}

\item{relative_humidity}{relative humidity in percent}
}
\value{
absolute humidity contained in 1 m^3 in l
}
\description{
Calculate the absolute humidity in 1 m^3 dependence of the temperature and the relative humidity
}
