% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/arcsine_trans.R
\name{arcsine_tran}
\alias{arcsine_tran}
\title{Arcsine transformation}
\usage{
arcsine_tran(y, n)
}
\arguments{
\item{y}{a numeric vector to be transformed.}

\item{n}{the number of elements in the numeric vector.}
}
\value{
the arcsine transformed vector.
}
\description{
This function performs an arcsine transformation.
}
\note{
The logit transformation has better proporties for transforming proportions.
}
\examples{
ntrials <- 40
successes <- 0.4
x <- rbinom(50, size = ntrials, prob = successes) / ntrials
arcsine_tran(x, ntrials)
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{arcsine}

