% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.numericSilent.R
\name{as.numericSilent}
\alias{as.numericSilent}
\title{Silent wrapper for coercing a vector to numeric}
\usage{
as.numericSilent(x)
}
\arguments{
\item{x}{vector of any type}
}
\value{
If \code{as.numeric(x)} produces an error or warning, \code{x} is
returned unchanged.  Otherwise, \code{as.numeric(x)} is returned.
}
\description{
Silent wrapper for coercing a vector to numeric
}
\examples{
as.numericSilent(c("this","that"))
as.numericSilent(c("2893.9","9423.48"))
as.numericSilent(c("392.1", "that"))
}
\author{
Landon Sego
}
\seealso{
\code{\link{as.numeric}}
}
\keyword{misc}

