% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelvar.R
\name{clear.labels}
\alias{clear.labels}
\title{Clearing labels created by Hmisc's package that sometimes cause troubles for other functions}
\usage{
clear.labels(x)
}
\description{
Clearing labels created by Hmisc's package that sometimes cause troubles for other functions
}
