% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{colend}
\alias{colend}
\title{Giving column names end with pattern}
\usage{
colend(df, pattern)
}
\arguments{
\item{df}{dataframes with column names containing string}

\item{pattern}{string pattern}
}
\description{
Giving column names end with pattern
}
\examples{
colend(mtcars$make, 'to')
}
\keyword{string}
