% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{colmatch}
\alias{colmatch}
\title{Giving column names match with pattern}
\usage{
colmatch(df, pattern)
}
\arguments{
\item{df}{dataframes with column names containing string}

\item{pattern}{string pattern}
}
\description{
Giving column names match with pattern
}
\examples{
colmatch(mtcars$make, 'to')
}
\keyword{string}
