% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{colMedians}
\alias{colMedians}
\title{Column Median}
\usage{
colMedians(x, ...)
}
\arguments{
\item{x}{dataframe or matrix}

\item{...}{additional arguments to be passed to \code{median()}}
}
\description{
The missing colMedian function. Works just like \code{colMeans()}
}

