% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of Variation}
\usage{
cv(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{additional arguments to be passed to \code{sd} and \code{mean}
i.e \code{na.rm = TRUE}}
}
\description{
Coefficient of variation
}

