% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/CVs.R
\name{CVs}
\alias{CVs}
\title{Coefficient of variation small samples}
\usage{
CVs(sd, m, n)
}
\arguments{
\item{sd}{the standard deviation of a variable.}

\item{m}{the mean of a variable.}

\item{n}{the number of elements in a variable.}
}
\value{
The small sample coefficient of variation.
}
\description{
This function calculates the coefficient of variation for small sample sizes.
}
\note{
The small sample coefficient of variation will converge on the large sample coefficient of variation as sample size increases.
}
\examples{
\dontrun{
CVs()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{variance}

