% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbb.R
\name{dbb}
\alias{dbb}
\alias{pbb}
\alias{qbb}
\alias{rbb}
\title{The Beta-Binomial Distribution}
\usage{
dbb(x, N, u, v)
pbb(q, N, u, v)
qbb(p, N, u, v)
rbb(n, N, u, v)
}
\arguments{
\item{x}{vector of qauntiles}

\item{N}{number of trials (a positive integer)}

\item{u}{first positive parameter of the beta distribution}

\item{v}{second positive parameter of the beta distribution}

\item{q}{vector of quantiles}

\item{p}{vector of probabilities}

\item{n}{number of observations}
}
\value{
\code{dbb} gives the density, \code{pbb} gives the distribution
function, \code{qbb} gives the quantile function, and \code{rbb} generates
random deviates.
}
\description{
Density, distribution function, quantile function, and random generation
for the beta-binomial distribution.  A beta-binomial variate
follows the binomial distribution with parameters
\code{N} and \code{p}, where the probability \code{p} of success itself has
a beta distribution with parameters \code{u} and \code{v}.
}
\details{
The beta-binomial distribution with parameters \emph{N}, \emph{u}, and
\emph{v} has mass function given by \emph{choose(N, x) * Beta(x + u, N - x + v) /
Beta(u, v)} for \emph{u > 0}, \emph{v > 0}, a positive integer \emph{N}, and
any nonnegative integer \emph{x}. Although one can express the integral in
closed form using generalized hypergeometric functions, the implementation
of distribution function used here simply relies on the cumulative sum of
the mass function.

The mean, \emph{mu}, and variance, \emph{sigma^2}, of the beta-binomial distribution
can be computed explicitly as \emph{mu = N * u / (u + v)} and
\emph{sigma^2 = (N * u * v * (N + u + v)) / ((u + v)^2 * (1 + u + v))}.

\code{dbb} and \code{qbb} do not recycle vector-valued arguments in
the usual fashion in R.  Arguments to these functions must have length 1
and/or one other common length. See example of \code{\link{pcbinom}} for more
information.

Non-integer values of \code{x} and \code{N} are rounded to the nearest
integer with a warning.
}
\section{Warning}{
 \code{dbb}, \code{pbb}, and \code{qbb} can be imprecise
when the probabilities for tail values are small.  This can result in
unusual behavior for \code{pbb} and \code{qbb}.
}
\examples{
# set up parameters
w <- 10
u <- 0.3 *w
v <- 0.7 * w
N <- 12

# generate random values from the beta-binomial
x <- rbb(1000, N, u, v)

# check that the empirical summary matches the theoretical one
summary(x)
qbb(c(0.25, 0.50, 0.75), N, u, v)

# check that the empirical histogram matches the theoretical density
hist(x, breaks=seq(-0.5, N + 0.55), prob = TRUE)
lines(0:N, dbb(0:N, N, u,v), type = "b")

# An example of the imprecision in pbb and qbb:  we would expect
# this to return 0:29
qbb(pbb(0:29, 29, 1, 33), 29, 1, 33)
}
\author{
Kevin R. Coombes, modified by Landon Sego
}
\references{
\url{http://bioinformatics.mdanderson.org/Software/OOMPA/Current/TailRank-manual.pdf}
}
\seealso{
\code{\link{dbeta}} for the beta distribution and
\code{\link{dbinom}} for the binomial distribution.
}
\keyword{misc}

