% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/densplot.R
\name{densplot}
\alias{densplot}
\title{Kernal density plot}
\usage{
densplot(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
This function creates a nice kernal density plot.
}
\examples{
x <- rnorm(100, 10, 3)
densplot(x)
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{density}}
}
\keyword{density}

