% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{dfbegin}
\alias{dfbegin}
\title{Giving df with column names begin with pattern}
\usage{
dfbegin(df, pattern)
}
\arguments{
\item{df}{dataframes with column names containing string}

\item{pattern}{string pattern}
}
\description{
Giving df with column names begin with pattern
}
\examples{
dfbegin(mtcars$make, 'to')
}
\keyword{string}
