% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{dfend}
\alias{dfend}
\title{Giving df with column names end with pattern}
\usage{
dfend(df, pattern)
}
\arguments{
\item{df}{dataframes with column names containing string}

\item{pattern}{string pattern}
}
\description{
Giving df with column names end with pattern
}
\examples{
dfend(mtcars$make, 'to')
}
\keyword{string}
