% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{dfmatch}
\alias{dfmatch}
\title{Giving df with column names match with pattern}
\usage{
dfmatch(df, pattern)
}
\arguments{
\item{df}{dataframes with column names containing string}

\item{pattern}{string pattern}
}
\description{
Giving df with column names match with pattern
}
\examples{
dfmatch(mtcars$make, 'to')
}
\keyword{string}
