% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_water.R
\name{diff_water}
\alias{diff_water}
\title{Calculate the difference in wawter volume given temperature and humidity inside and outside}
\usage{
diff_water(
  temperature_outside,
  relative_humidity_outside,
  temperature_inside,
  relative_humidity_inside,
  volume = 300
)
}
\arguments{
\item{temperature_outside}{temperature outside in degress celsius}

\item{relative_humidity_outside}{relative humidity outside in percent}

\item{temperature_inside}{temperature inside in degress celsius}

\item{relative_humidity_inside}{relative humidity inside in percent}

\item{volume}{Volume in cubic meters that is exchanged}
}
\value{
difference in water content before and after changing the air
}
\description{
Calculate the difference in wawter volume given temperature and humidity inside and outside
}
\examples{
diff_water(8, 80, 19, 70)
}
