% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2character.R
\name{factor2character}
\alias{factor2character}
\title{Converts all factor variables in a dataframe to character variables}
\usage{
factor2character(dframe)
}
\arguments{
\item{dframe}{A dataframe}
}
\value{
The same dataframe with the factor variables converted to character
variables.
}
\description{
Converts all factor variables in a dataframe to character variables
}
\examples{

x <- data.frame(a=factor(c(rep(1,4),rep(2,4),rep(3,4))), y=rnorm(12))
str(x)
x <- factor2character(x)
str(x)

}
\author{
Landon Sego
}
\keyword{manip}
\keyword{misc}
