% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2num.R
\name{factor2num}
\alias{factor2num}
\title{A simple function for converting factors to numeric values}
\usage{
factor2num(x)
}
\arguments{
\item{x}{A vector of type \code{factor}}
}
\value{
\code{x} converted to a numeric vector
}
\description{
A simple function for converting factors to numeric values
}
\examples{
# Define a factor object
y <- factor(5:7)
y

# incorrectly convert to numeric
as.numeric(y)

# correctly convert
factor2num(y)

}

