% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{signal}
\alias{signal}
\title{Generic for \code{signal}}
\usage{
signal(object, ...)
}
\arguments{
\item{object}{The object on which the generic operates}

\item{\dots}{Arguments passed to specific methods}
}
\description{
Generic for \code{signal}
}
