% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_polarhist.R
\name{geom_polar_hist}
\alias{geom_polar_hist}
\title{polar histogram}
\usage{
geom_polar_hist(x, group)
}
\arguments{
\item{x}{data}

\item{group}{group}
}
\description{
ggplot polar histogram set from 0 to 360 degrees
}

