% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtension.R
\name{getExtension}
\alias{getExtension}
\title{Get the extension of a vector of filenames}
\usage{
getExtension(vec, split.char = ".")
}
\arguments{
\item{vec}{Character vector (usually containing filenames)}

\item{split.char}{A single character used to split the character strings}
}
\value{
Character vector of filename extensions
}
\description{
Get the extension of a vector of filenames, assuming that the extension is the set of characters that
follows the last \code{"."}.  A wrapper for \code{\link{grabLast}}.
}
\details{
Assumes paths are delineated using forward slashes.  If an \code{NA} is
supplied, then an \code{NA} is returned.  If the desired string doesn't
exist (see examples below), a \code{""} is returned.
}
\examples{
getExtension(c(a = "this old file.doc",
               b = "that young file.rtf",
               c = "this.good.file.doc",
               d = "this_bad_file",
               e = "thisfile.",
               f = NA,
               g = "that.this.pdf",
               h = ".", i = ""))

# An example with 'real' files
files <- dir(file.path(path.package(package = "Smisc"), "data"), full.names = TRUE)
print(files)
getExtension(files)
}
\seealso{
Additional functions for filename manipulations:  \code{\link{stripExtension}}, \code{\link{stripPath}},
\code{\link{getPath}}, \code{\link{grabLast}}, \code{\link{basename}}, \code{\link{dirname}}
}
\author{
Landon Sego
}
\keyword{misc}
