% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envunction.R
\name{group_hist}
\alias{group_hist}
\title{Grouped histogram}
\usage{
group_hist(dataframe, group, vec)
}
\description{
Histogram density plots (in percentage, not counts) of a variable grouped by another variable
}
