% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hpd}
\alias{hpd}
\alias{plot.hpd}
\alias{print.hpd}
\title{Calculate the highest posterior density credible interval for a unimodal density}
\usage{
hpd(pdf, support, prob = 0.95, cdf = NULL)

\method{print}{hpd}(hpdObject)

\method{plot}{hpd}(hpdObject, ...)
}
\arguments{
\item{pdf}{Function that takes a single numeric vector argument that returns a vector
of probability density values}

\item{support}{A numeric vector of length 2 giving the interval over which the random variable has
support (i.e. for which the pdf is positive).  For now, this must be a finite interval.
Intervals for random variables within infinite support
can still be calculated by setting the values of support to be suitably large and/or small.  See examples.}

\item{prob}{A numeric value in (0, 1] indicating the size of the desired probability for the credible
interval.}

\item{cdf}{A function that takes a single (not necessarily vector) argument and returns the cumulative probability.
If \code{NULL}, the pdf is integrated as needed to calculate probabilities
as needed.  However, when possible, it's best to provide the cdf.}

\item{hpdObject}{object of class \code{hpd}, returned by \code{hpd}}

\item{\dots}{Additional agruments that may be passed to \code{\link{plot.default}} or \code{\link{abline}}}
}
\value{
A list of class \code{hpd} that contains the following elements:
\describe{
\item{lower}{The lower endpoint of the highest posterior density interval}
\item{uper}{The lower endpoint of the highest posterior density interval}
\item{prob}{The acheived probability of the interval}
\item{cut}{The horizontal cut point that gave rise to the interval}
\item{mode}{The mode of the density}
\item{pdf}{The probability density function}
\item{support}{The support of the pdf}
}
}
\description{
Calculate the highest posterior density credible interval for a unimodal density
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Prints the lower and upper limits of the credible interval, along with the achieved
probabilty of that interval.

\item \code{plot}: Plots the density, overlaying the lower and upper limits of the credible interval
}}
\examples{
# A credible interval using the standard normal
int <- hpd(dnorm, c(-5,5), prob = 0.90)
print(int)
plot(int)

# The gamma density
int <- hpd(function(x) dgamma(x, shape = 2, rate = 0.5), c(0, 20),
           cdf = function(x) pgamma(x, shape = 2, rate = 0.5), prob = 0.8)
print(int)
plot(int)

# A credible interval using the Beta density
dens <- function(x) dbeta(x, 7, 12)
dist <- function(x) pbeta(x, 7, 12)
int <- hpd(dens, c(0, 1), cdf = dist)
print(int)
plot(int)
}
\author{
Landon Sego
}

