% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ifelse1}
\alias{ifelse1}
\title{Non-vectorized version of ifelse}
\usage{
ifelse1(test, yes, no)
}
\arguments{
\item{test}{An expression that resolves to logical of length 1, i.e.,
\code{TRUE} or \code{FALSE}}

\item{yes}{An object that is returned if \code{test} resolves to \code{TRUE}}

\item{no}{An object that is returned if \code{test} resolves to \code{FALSE}}
}
\value{
Either \code{yes} or \code{no}, depending on \code{test}.
}
\description{
Non-vectorized version of \code{\link{ifelse}}.
}
\examples{
ifelse1(7 == 7, rnorm(5), letters[1:5])
ifelse1(7 == 2, rnorm(5), letters[1:5])
}
\author{
Landon Sego
}
\seealso{
\code{\link{ifelse}}
}

