% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{ifelse2}
\alias{ifelse2}
\title{ifelse but treating NA as FALSE}
\usage{
ifelse2(condition, then, otherwise)
}
\arguments{
\item{condition}{condition for recoding}

\item{then}{value if condition is true}

\item{otherwise}{value if condition is false or NA due to missing values in vector which condition base}
}
\description{
ifelse but treating NA as FALSE
}
\examples{
z = c(1,2,NA,3)
y = c(1,NA,2,3)
ifelse(y<2,'T','F')  # [1] "T" NA  "F" "F"
ifelse2(y<2,'T','F') # [1] "T" "F" "F" "F"
}
\keyword{logical}
