% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ipak.R
\name{ipak}
\alias{ipak}
\title{Multiple package install and load}
\usage{
ipak(pkg)
}
\arguments{
\item{pkg}{a vector of package names in quotes.}
}
\description{
This function installs new packages and loads them into an R session.
}
\examples{
\dontrun{
packages <- c("ggplot2", "scales", "plyr")
ipak(packages)
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{install.packages}}
}
\keyword{package}

