% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_integer.R
\name{is_integer}
\alias{is_integer}
\title{What is.integer() should be}
\usage{
is_integer(x)
}
\arguments{
\item{x}{number of vector of numbers}
}
\value{
boolean
}
\description{
Determines if a number is an integer.
}

