% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda}
\alias{lambda}
\title{Pagel's lambda}
\usage{
lambda(model)
}
\arguments{
\item{model}{a model object. Accepts models of class MCMCglmm.}

\item{gen}{generalized version of lambda including all random effects in denominator? Defaults to FALSE.}
}
\value{
a vector of Pagel's lambda values whose length equals the size of the posterior sample.
}
\description{
This function calculates Pagel's lambda parameter. For models of class MCMCglmm this is also called H^2.
}
\examples{
\dontrun{
lambda_values <- lambda(model1)
plot(density(lambda_values))
posterior.mode(lambda_values)
mean(lambda_values)
median(lambda_values)
HPDinterval(lambda_values)
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\references{
Pagel (1999)
}
\keyword{lambda}

