% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/last.R
\name{last}
\alias{last}
\title{Return last element(s) of vector}
\usage{
last(x)
}
\arguments{
\item{x}{a vector or matrix.}

\item{n}{number of elements to return.}
}
\description{
This function returns the last element(s) of a vector. Use last(x) instead of x[length(x)]. Works on matrices too.
}
\examples{
x <- letters
last(x)
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{tail}}
}
\keyword{tail}

