% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_dataframe.R
\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{Converts list to a dataframe in a sensible way}
\usage{
list_to_dataframe(l, col_name = NULL)
}
\arguments{
\item{l}{list of dataframes to be converted into single dataframe}

\item{col_name}{(optional) name of column to put element names under}
}
\value{
dataframe
}
\description{
Given a list of dataframes with the same columns, this function will row bind
them together, and if passed a \code{col_name} arguement, will produce a
column containing their original element name
}
\examples{
split_iris <- split(iris, iris$Species)
list_out <- lapply(split_iris, function(x) {x['Species'] <- NULL; x})
list_to_dataframe(list_out, col_name = "Species")
}

