% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_csv.R
\name{load_csv}
\alias{load_csv}
\title{load all csv files in a folder}
\usage{
load_csv(path)
}
\arguments{
\item{path}{string, path to folder}
}
\description{
Given a path to a folder, this function will load all files ending with
.csv into the global namespace
}

