% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadObject}
\alias{loadObject}
\title{Loads and returns the first object in an ".Rdata" file}
\usage{
loadObject(Rdata.file)
}
\arguments{
\item{Rdata.file}{A character string containing the '.Rdata' filename}
}
\value{
The first object in the '.Rdata' file.
}
\description{
Loads and returns the first object in the ".Rdata" file.  This is useful
for naming the object in the ".Rdata" something other than the name it was
saved as.
}
\details{
If the '.Rdata' file contains more than one object, only the first object
is returned and a warning is produced.
}
\examples{
# Make an example object
x <- data.frame(a=1:10, b=rnorm(10))

# Create a unique filename with a time stamp
# imbedded in the filename
file.name <- timeStamp("demo_load_object", "Rdata")

# Save and reload the file
save(x, file=file.name)
y <- loadObject(file.name)
print(y)

# Delete the saved object
unlink(file.name)
}
\author{
Landon Sego
}
\keyword{misc}

