% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{locate}
\alias{locate}
\title{Locating rows containing string}
\usage{
locate(string, pattern)
}
\arguments{
\item{string}{vectors containing string}

\item{pattern}{string pattern}
}
\description{
This function locating row numbers that contain string pattern.
}
\examples{
locate(mtcars$make, 'toyota | honda')
}
\keyword{string}
