% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{List of objects}
\usage{
lsos(..., n = 10)
}
\arguments{
\item{...}{an (optional) object name.}

\item{n}{the number of objects to return.}
}
\description{
This function is an improved list of objects in the R global environment.
}
\examples{
\dontrun{
lsos()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\references{
taken from: \url{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
}
\seealso{
\code{\link{ls}}
}
\keyword{ls}

