% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/merge_all.R
\name{merge_all}
\alias{merge_all}
\title{Multiple merges}
\usage{
merge_all(dflist, by.var)
}
\arguments{
\item{dflist}{a list of data frames.}

\item{by.var}{a vector of variable names in quotes to use in merging the data frames.}
}
\value{
a merged data frame.
}
\description{
This function merges multiple data frames.
}
\examples{
\dontrun{
merge_all()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{merge}}
}
\keyword{merge}

