% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{mixreg}
\alias{mixreg}
\title{Caution: this function fails to converge if prior precision is set too high (>5). Better set diffuse prior.}
\usage{
mixreg(nrep, y, x, n_group, priorp, priorbeta, priorprec, a1, a2, P, betas, sig,
  tau)
}
\description{
Caution: this function fails to converge if prior precision is set too high (>5). Better set diffuse prior.
}
