% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/notin.R
\name{notin}
\alias{\%!in\%}
\alias{notin}
\title{Inverse Value Matching}
\usage{
x \%!in\% y
}
\value{
A logical vector, indicating if a match was NOT located for each element of x: thus the values are TRUE or FALSE and never NA.
}
\description{
Complement of \code{\%in\%}. Returns the elements of \code{x} that are not in \code{y}.
}
\examples{
x <- letters[1:5]
y <- letters[4:8]
x \%!in\% y
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{\%in\%}}
}
\keyword{match}

