% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/packinfo.R
\name{packinfo}
\alias{packinfo}
\title{List installed packages and versions}
\usage{
packinfo()
}
\description{
This function lists installed packages and their versions.
}
\examples{
packinfo()
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{installed.packages}}
}
\keyword{package}

