% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{PCbiplot}
\alias{PCbiplot}
\title{Biplot from a principal components analysis (PCA)}
\usage{
PCbiplot(PC, d, colors, legend_t, varnames, labels, title, x = "PC1",
  y = "PC2")
}
\arguments{
\item{PC}{a prcomp object obtained by conducting a principal components
analysis.}

\item{d}{a data frame of metadata containing the sample identifiers and
categorical classification of the samples.}

\item{colors}{a character vector of same length as the number of levels in
the classfication.}

\item{legend_t}{the name of the column in d that gives the sample
classification.}

\item{varnames}{a character vector of names for the variables in the biplot.}

\item{labels}{a logical indicating whether samples should be labelled.}

\item{title}{a title for the biplot.}

\item{x}{principal component to be plotted along the horizontal (x) axis or
abcissa.}

\item{y}{principal component to be plotted along the vertical (y) axis or
ordinate.}
}
\value{
A ggplot2 object
}
\description{
Biplot from a principal components analysis (PCA)
}
