% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pctin.R
\name{pctin}
\alias{\%pctin\%}
\alias{pctin}
\title{percent in}
\usage{
x \%pctin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched}

\item{table}{vector or NULL: the values to be matched against}
}
\value{
percentage of elements of \code{x} that are in \code{table}
}
\description{
calculate the percentage of elements of \code{table} that are in \code{x}
}
\examples{
letters[1:10] \%pctin\% letters[1:3] # 30\% of the second arg ar in the first
}
\author{
gsee
}

