% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{\%+\%}
\alias{plus}
\title{Concatenate two strings}
\usage{
x \%+\% y
}
\arguments{
\item{x}{a character string.}

\item{y}{a character string.}
}
\value{
Same as paste0(x, y).
}
\description{
Concatenates two strings into a single string.
}
\examples{
x <- letters[1:5]
y <- letters[4:8]
x \%+\% y
}

