\name{pnlStat}
\alias{pnlStat}
\docType{package}
\title{
A collection of miscellaneous functions
}
\description{
Collection of convenient functions used in the PNNL Stat Group
}
\details{
\tabular{ll}{
Package: \tab pnlStat\cr
Type: \tab Package\cr
Version: \tab 2013.03.12\cr
Date: \tab 2013-03-12\cr
License: \tab Unknown \cr
}
}
\author{
Members of the PNNL Statistical Sciences group

Maintainer: Landon Sego <Landon.Sego@pnl.gov>
}
% \references{
% ~~ Literature or other references for background information ~~
% }
\keyword{package}
% ~~ Optionally other keywords from doc/KEYWORDS, one per line
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
