% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerData.R
\docType{data}
\name{PowerData}
\alias{PowerData}
\title{An example of power data}
\format{The format is: Named num [1:12] 5.15 5.15 5.15 5.16 5.16 ...  -
attr(*, "names")= chr [1:12] "2008-05-06 17:00:01" "2008-05-06 17:00:17"
"2008-05-06 17:00:21" "2008-05-06 17:00:32" ...}
\description{
A small, altered, subset of total rack power for one of the racks in a data
center.  Used for illustration of \code{\link{timeIntegration}}
}
\examples{

data(PowerData)
print(PowerData)
rm(PowerData, envir=.GlobalEnv)

}
\keyword{datasets}
