% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{ptab}
\alias{ptab}
\title{Producing proportion table easily}
\usage{
ptab(..., margin = NULL, round = 1)
}
\arguments{
\item{data}{the data to make cross table}

\item{by}{whether the proportion is calculated with regard to row total or column total}
}
\description{
Producing proportion table easily
}
\examples{
ptab(df \%>\% select(x,y), by=2)
}
\keyword{cross}
\keyword{tabulation}
