% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_all.R
\name{rbind_all}
\alias{rbind_all}
\title{opens multiple csv files and binds them together}
\usage{
rbind_all(path, col_name = "origin", ...)
}
\arguments{
\item{path}{string, path to folder containing .csv files}

\item{col_name}{string, name to call new column listing the origin for
each file.}

\item{...}{additional arguments to \code{read.csv}}
}
\description{
opens all csv files in a folder, and assigns each an identifying column
indicating which file it came from, and binds all data frames together
in a row-wise fashion. Data frames have to have the same columns and column headers
}

