% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{recode_eq}
\alias{recode_eq}
\title{recoding from the old value to new value}
\usage{
recode_eq(data, varvector, old.value, new.value)
}
\arguments{
\item{data}{is the dataframe}

\item{varvector}{is a vector with string elements of variable names.}

\item{old.value}{is the value which will be recoded}

\item{new.value}{is the value which will become the replacement}

\item{lowbound}{is the lower bound of values, whatever above that bound will be recoded}

\item{upperbound}{is the upper bound of values, whatever below that bound will be recoded}

\item{not.value}{is the value which whatever that is not that value will be recoded}
}
\description{
recoding from the old value to new value
}
\examples{
new.data <- recoding(old.data,c('v1','v2'), 3,0)
new.data <- recode.na(old.data,c('v1','v2'), 96)
}
\keyword{recode}
