% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{recode_ie}
\alias{recode_ie}
\title{recoding based on condition, otherwise keep the original value}
\usage{
recode_ie(data, condition, value)
}
\arguments{
\item{data}{vector to recode}

\item{condition}{condition for recoding}

\item{value}{value to put on vector}
}
\description{
recoding based on condition, otherwise keep the original value
}
\examples{
z = c(1,2,NA,3)
y = c(1,NA,2,3)
z = recode2(z,y<=2,10)
z # [1] 10  2 10  3
}
\keyword{recode}
