% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{recodestring}
\alias{recodestring}
\title{recoding based on rows containing string}
\usage{
recodestring(x, strings, pattern, value)
}
\arguments{
\item{x}{vector to recode}

\item{pattern}{string pattern}

\item{value}{value to put on vector}

\item{string}{vectors containing strings}
}
\description{
This function locating row numbers that contain string pattern.
}
\examples{
recodestring(mtcars$japan, mtcars$make, 'toyota | honda', 1)
}
\keyword{string}
