% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_string.R
\name{reverse_str}
\alias{reverse_str}
\title{Reverse string}
\usage{
reverse_str(s)
}
\arguments{
\item{s}{string}
}
\value{
string. reverse of s
}
\description{
As reverse("string") doesn't work, here's the function that does
}
\examples{
reverse_str("string")
}

