% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_col_na.R
\name{rm_col_na}
\alias{rm_col_na}
\title{Removes columns consisting of only NA}
\usage{
rm_col_na(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Given a dataframe this will remove the columns that contain only NA values.
}

