\name{rma}
\alias{rma}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
 Remove all objects from the global environment
}
\description{
 Remove all objects from the global environment
}
\usage{
rma()
}
%- maybe also 'usage' for other objects documented here.
\details{
Alias for \code{rm(list=ls())}.
}
%\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
%}
%\references{
%% ~put references to the literature/web site here ~
%}
\author{
Landon Sego
}
%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
