% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{rmvn}
\alias{rmvn}
\title{rmvn is a function to take random values from multivariate normal distribution with 'n' number of observation, mean 'u', and variance-covariance matrix 'sigma'}
\usage{
rmvn(n = 1, u = 0, sigma = 1)
}
\description{
rmvn is a function to take random values from multivariate normal distribution with 'n' number of observation, mean 'u', and variance-covariance matrix 'sigma'
}
