% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/robSE.R
\name{robSE}
\alias{robSE}
\title{Robust standard errors for OLS models}
\usage{
robSE(model, type = "HC4m")
}
\arguments{
\item{model}{an object of class lm.}

\item{type}{the type of robust standard error to calculate. Defaults to "HC4m".}
}
\description{
This function returns robust standard errors for objects of class lm.
}
\examples{
\dontrun{
robSE()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{robust}

