% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/roundup.R
\name{roundup}
\alias{roundup}
\title{Round floats up}
\usage{
roundup(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\description{
This function redefines the round function to "round up" instead of "round to even".
}
\examples{
x <- rnorm(10, 10, 3)
roundup(x)
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{round}}
}
\keyword{round}

