% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{rowMedians}
\alias{rowMedians}
\title{Row Median}
\usage{
rowMedians(x, ...)
}
\arguments{
\item{x}{dataframe or matrix}

\item{...}{additional arguments to be passed to \code{median()}}
}
\description{
The missing rowMedian function. Works just like \code{rowMeans()}
}

