% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample_df.R
\name{sample_df}
\alias{sample_df}
\title{Sample a data frame}
\usage{
sample_df(df, n = 3, ordered = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{n}{the number of lines to sample.}

\item{ordered}{order the rows? Defaults to TRUE.}
}
\description{
This function randomly samples rows from a dataframe.
}
\examples{
\dontrun{
sample_df()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{sample}

