% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectElements.R
\name{selectElements}
\alias{selectElements}
\title{Validate selected elements from a character vector}
\usage{
selectElements(elements, cVec)
}
\arguments{
\item{elements}{elements to select from \code{cVec}.  Can be a logical vector, a vector of 1's and 0's (that is
converted to Boolean), a vector of numeric indexes, or a character vector of column names.
If \code{elements == NULL}, \code{NULL} is returned.}

\item{cVec}{A character vector from which to select elements}
}
\value{
A character vector with elements that were selected from \code{cVec} using \code{elements}
}
\description{
Validate selected elements from a character vector using a variety of selection mechanisms:  logical, names, or
numerical indexes
}
\details{
This function is useful for selecting rows or columns from data frames, while providing informative error messages
if the \code{elements} for selection are specified incorrectly.  Note that logical vectors (or binary vectors of 0's
and 1's) are not recycled as usual: they must be the same length as \code{cVec}.
}
\examples{
# Define the column names
cnames <- letters[1:5]
cnames

# Select the 1st and 3rd column names using a variety of approaches
selectElements(c("a", "c"), cnames)
selectElements(c(1, 3), cnames)
selectElements(c(TRUE, FALSE, TRUE, FALSE, FALSE), cnames)
selectElements(c(1, 0, 1, 0, 0), cnames)

# If you don't want to select any of them
selectElements(NULL, cnames)
}

