% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_var.R
\name{show_var}
\alias{show_var}
\title{Show the name and the value of a variable}
\usage{
show_var(x)
}
\arguments{
\item{x}{variable to be shown}
}
\value{
no return value
}
\description{
Can be used for debugging purposes
}
\examples{
x <- 10
show_var(x)
}
