% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Smisc-package.R
\docType{package}
\name{Smisc-package}
\alias{Smisc}
\alias{Smisc-package}
\title{Sego Miscellaneous}
\description{
Sego Miscellaneous:  A collection of functions for statistical computing and data manipulation
}
\details{
\tabular{ll}{
Package: \tab Smisc\cr
Type: \tab Package\cr
Version: \tab 0.2.9\cr
Date: \tab 2015-06-17\cr
License: \tab LICENSE\cr
}
}
\examples{
help(package = Smisc)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

