% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Smisc}
\alias{Smisc}
\alias{Smisc-package}
\title{A collection of miscellaneous functions}
\description{
Collection of convenient functions
}
\details{
\tabular{ll}{ Package: \tab Smisc\cr Type: \tab Package\cr Version: \tab
2013.03.12\cr Date: \tab 2013-03-12\cr License: \tab Unknown \cr }
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

