% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sortDF}
\alias{sortDF}
\title{Sorts a dataframe by one or more variables}
\usage{
sortDF(form, dat)
sortDF(dat, form)
}
\arguments{
\item{form}{A formula with the variable names to use for sorting}

\item{dat}{The dataframe to sort}
}
\value{
The sorted dataframe
}
\description{
Sorts a dataframe by one or more variables, ascending or descending
 The order of the arguments \code{form} and \code{dat} is
  interchangeable.  In the formula, use + for ascending, - for
  decending. Sorting is left to right in the formula.
  For example, to sort the dataframe \code{Oats} with sorting variables \code{Block} and
  \code{Variety}, useage is either of the following:
}
\details{
\code{sortDF(~ Block - Variety, Oats)}  or

\code{sortDF(Oats, ~ - Variety + Block)}
}
\examples{
d <- data.frame(b = factor(c("Hi","Med","Hi","Low"), levels = c("Low","Med","Hi"),
                           ordered = TRUE),
                x = c("A", "D", "A", "C"),
                y = c(8, 3, 9, 9),
                z = c(1, 1, 1, 2))

# Sort by descending z, descending b
sortDF(~ -z - b, d)

# Sort by ascending x, ascending y, and ascending z
sortDF(~x + y + z, d)

# Sort by descending x, ascending y, ascending z
sortDF(~ -x + y + z, d)

# Sort by ascending x, descending y, ascending z
sortDF(d, ~x - y + z)
}
\author{
Kevin Wright with ideas from Any Liaw and small edits by Landon Sego
}

