% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/source_https.R
\name{source_https}
\alias{source_https}
\title{Source R script over HTTPS}
\usage{
source_https(url, ...)
}
\arguments{
\item{url}{a valid URL.}

\item{...}{additional arguments.}
}
\description{
This function sources an R script that is hosted over a HTTPS connection (e.g., github).
}
\examples{
\dontrun{
source_https()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{source}}
}
\keyword{source}

