% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envunction.R
\name{spready}
\alias{spready}
\title{make dataset spread-ready, useful to inspect problem in spreading}
\usage{
spready(data, item, ...)
}
\description{
make dataset spread-ready, useful to inspect problem in spreading
}
