% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a numeric argument}
\usage{
square(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
This function yields the square of a numeric argument
}
\examples{
square(7)
}
