% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstable.R
\name{ss_xtabs}
\alias{ss_xtabs}
\title{A crosstable in LaTeX}
\usage{
ss_xtabs(formula, data, caption = "", addmargins = NULL, digits = 0)
}
\arguments{
\item{formula}{a formula involving columns names of data. This formula will
produce the crosstable and will be passed to xtabs function from stats
package.}

\item{data}{a dataframe}

\item{caption}{caption for the table produced by the function.}

\item{addmargins}{parameter for addmargins from stats package.}

\item{digits}{number of digits in the output.

The following parameters will be passed to xtabs and addmargins from stats
    package.}
}
\value{
a latex crosstable with caption and margins
}
\description{
A crosstable in LaTeX
}
