% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopifnotMsg.R
\name{stopifnotMsg}
\alias{stopifnotMsg}
\title{Check multiple conditions and return coresponding error messages}
\usage{
stopifnotMsg(...)
}
\arguments{
\item{\dots}{Pairs of logical conditions and error messages. See Examples.}
}
\value{
A call to \code{\link{stop}} with the error messages from each condition that was \code{FALSE}.  If all
conditions are \code{TRUE}, returns \code{NULL}.
}
\description{
A more flexible version of \code{\link{stopifnot}} that allows you to control the error message returned by
each condition that doesn't test true
}
\examples{
# A simple function
aFunction <- function(x, a = 10, b = "text") {

  # Check the arguments of the function
  stopifnotMsg(is.numeric(x),   "'x' must be numeric",
               is.numeric(a),   "'a' must be numeric",
               a > 9,           "'a' must be 10 or more",
               is.character(b), "'b' must be character")

  return(paste(x, a, b, sep = " -- "))

}

# This should run without error
aFunction(12, a = 13, b = "new")

\dontrun{
# This should produce an error with 3 messages:
aFunction("new", a = 7, b = 5)

# And this should produce an error with 1 message:
aFunction(33, a = "bad")
}
}
\author{
Landon Sego
}

