% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize}
\alias{summarize}
\title{Summary of the data set}
\usage{
summarize(obj, name = "")
}
\arguments{
\item{obj}{a dataframe.}

\item{name}{name of the dataset.}
}
\description{
The summary produced by this function is the one from Hmisc package.
Here we have isolated the function and used it to render a LaTeX code that
can be used to produce PDF files.
}
