% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Summary plots}
\usage{
summary_plot(df, aspect = 1)
}
\arguments{
\item{df}{a data frame.}

\item{aspect}{the aspect ratio. Defaults to 1.}
}
\description{
This function produces summary histograms for each variable in a data frame. It splits the screen and finds the optimal number of columns and rows to be as close as possible to the desired aspect ratio.
}
\examples{
\dontrun{
summary_plot()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{summary}

