% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summary_text.R
\name{summary_text}
\alias{summary_text}
\title{Summary statistics}
\usage{
summary_text(df)
}
\arguments{
\item{df}{a data frame.}
}
\description{
This function produces summary statistics.
}
\examples{
\dontrun{
summary_text()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\seealso{
\code{\link{summary}}
}
\keyword{summary}

