% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summaryMCMC.R
\name{summaryMCMC}
\alias{summaryMCMC}
\title{Summary of MCMCglmm model}
\usage{
summaryMCMC(model, prob = 0.95, m = 2, digits = 4)
}
\arguments{
\item{model}{a model of class MCMCglmm.}

\item{prob}{the range for the highest probability density interval. Defaults to 95\%.}

\item{m}{the magnitude.}

\item{digits}{the number of digits to round to.}
}
\description{
This function produces a quick summary of a MCMCglmm model.
}
\examples{
\dontrun{
summaryMCMC()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\references{
Forked from Schaun Wheeler's GitHub.
}
\keyword{summary}

