% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summary standard errors and confidence intervals}
\usage{
summarySE(df = NULL, measurevar, groupvars = NULL, na.remove = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{measurevar}{the name of a column that contains the variable to be summariezed.}

\item{groupvars}{a vector containing names of columns that contain grouping variables.}

\item{na.remove}{should missing values be removed? Defaults to FALSE.}

\item{conf.interval}{the range of the confidence interval. Defaults to 95\%.}

\item{.drop}{drop redundant factor levels? Defaults to TRUE.}
}
\description{
This function calculates count, mean, standard error of the mean, and confidence intervals of the mean for grouped variables.
}
\examples{
\dontrun{
summarySE()
}
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{summary}

