% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_v.R
\name{switch_v}
\alias{switch_v}
\title{Vectorized version of the \command{switch} command}
\usage{
switch_v(EXPR, ...)
}
\arguments{
\item{EXPR}{bla}

\item{...}{bla}
}
\value{
absolute humidity contained in 1 m^3 in l
}
\description{
Vectorized version of the \command{switch} command
}
\examples{
x <- rep(c('a', 'b', 'c'), each = 3)
switch_v(x, 
         a = 'first', 
         b = 'second',
         'other')
}
