% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{timeData}
\alias{timeData}
\title{Generic data frame with a time variable}
\format{A data frame with 414 observations on the following 2 variables.
\describe{ \item{list("time")}{a POSIXt} \item{list("x")}{a numeric vector}
}}
\description{
Generic data frame with a time variable to support the example in
\code{\link{smartTimeAxis}}.
}
\examples{
data(timeData)
}
\keyword{datasets}

