% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeDiff.eg.R
\docType{data}
\name{timeDiff.eg}
\alias{timeDiff.eg}
\title{Four short time series}
\description{
Four short times series, stored in a list, for use in the \code{\link{timeDiff}} example.
}
\examples{

data(timeDiff.eg)

}
\keyword{datasets}
