% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/timeStamp.R
\name{timeStamp}
\alias{timeStamp}
\title{Imbeds the present datetime into a file name}
\usage{
timeStamp(description, extension)
}
\arguments{
\item{description}{Character string giving the base name of the file}

\item{extension}{Character string giving the extension of the file
(excluding the period)}
}
\value{
Character string of the form
\code{description_YYYY-MM-DD_HHMMSS.extension}
}
\description{
Imbeds the present datetime into a file name
}
\examples{
timeStamp("aFilename", "txt")
}
\author{
Landon Sego
}
\keyword{misc}

