% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim whitespace}
\usage{
trim(x, leading = TRUE, trailing = TRUE)
}
\arguments{
\item{x}{a character string or character vector.}

\item{leading}{trim leading whitespace? Defaults to TRUE.}

\item{trailing}{trim trailing whitespace? Defaults to TRUE.}
}
\description{
This function returns strings w/o leading and/or trailing whitespace.
}
\examples{
x <- c(" help", "me ", " to ", "remove", "whitespace  ")
trim(x)
}
\author{
Steven Worthington \email{sworthington@iq.harvard.edu}
}
\keyword{whitespace}

