% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validEmailAddress.R
\name{validEmailAddress}
\alias{validEmailAddress}
\title{Indicates whether a vector of email addreses may be valid}
\usage{
validEmailAddress(vec)
}
\arguments{
\item{vec}{Character vector of email addresses}
}
\value{
Logical vector which is \code{TRUE} when the email address is
valid.
}
\description{
Indicates whether a vector of email addreses may be valid
}
\details{
Checks the following conditions:

1: Only 1 @ allowed

2: Should have text on both sides of the @

3: Text on right hand side (RHS) of @ should have at least one dot

4: RHS text should not end with a dot

5: The suffix of the RHS should have 2 or 3 characters (e.g. 'com', 'net',
'uk')

6: RHS should not begin with a dot

7: No illegal characters--however, it does't check for a backslash
}
\examples{
validEmailAddress(c(NA, "@","this@that","this@that@that","this@that.com.ar",
                      "@this.com","that@","this@that.","this@.that","this@.that.com",
                       "this@that.com","this(@that.com","this@that .com"))
}
\author{
Landon Sego
}
\references{
Based in part on
\url{http://groups.google.com/group/eaut/web/rules-for-valid-email-addresses}
}
\keyword{misc}

