% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{vbegin}
\alias{vbegin}
\title{Giving vector of string begin with pattern}
\usage{
vbegin(string, pattern)
}
\arguments{
\item{string}{vectors containing string}

\item{pattern}{string pattern}
}
\description{
Giving vector of string begin with pattern
}
\examples{
begin(mtcars$make, 'to')
}
\keyword{string}
