% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{vend}
\alias{vend}
\title{Giving vector of string end with pattern}
\usage{
vend(string, pattern)
}
\arguments{
\item{string}{vectors containing string}

\item{pattern}{string pattern}
}
\description{
Giving vector of string end with pattern
}
\examples{
end(mtcars$make, 'to')
}
\keyword{string}
