% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelvar.R
\name{write_dta}
\alias{write_dta}
\title{Wrapper for write.dta}
\usage{
write_dta(dataframe, file, labels = NULL, ...)
}
\arguments{
\item{labels}{vector with same length as number of columns, the default is using variables' label attributes that comes from Hmisc package}
}
\description{
Piping-ready and easy labelling variables in the resulted stata file
}
\examples{
write_dta(mtcars, file = '~/test.dta') # use Hmisc's variable labels
mtcars \%>>\% write_dta(file = '~/test.dta') # piping-ready, equivalent with above
write_dta(mtcars, file = '~/test.dta', labels = labelvar) # use own-supplied label vector
}
