% Generated by roxygen2 (4.0.1): do not edit by hand
\name{comboList}
\alias{comboList}
\title{Produces all possible combinations of a set of linear model predictors}
\usage{
comboList(n.pred, outFile = NULL, njobs = 1, verbose = FALSE)
}
\arguments{
\item{n.pred}{integer indicating the number of predictors}

\item{outFile}{text string indicating the .Rdata file to which the returned
list of predictor combinations will be saved.  If NULL, then no file is
saved.}

\item{njobs}{Integer indicating the number of parallel jobs to be used in
calculating the combinations, using \code{\link{mclapply}}}

\item{verbose}{\code{= TRUE} indicates that a message regarding the saving
of the output file will be printed}
}
\value{
A list of class \code{combolist} is invisibly returned with the two
components shown below.  If \code{outFile} is not \code{NULL}, this same
list is saved to \code{outFile}: \item{len}{The total number of
combinations} \item{pList}{A list where each element contains an integer
representation of one combination of the predictors}
}
\description{
Produces a list representing all possible combinations of linear model
predictors
}
\details{
Uses \code{\link{combn}} to identify the combinations.
}
\examples{
x <- comboList(3)
print(x)
}
\author{
Landon Sego
}
\keyword{misc}

