% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_water.R
\name{diff_water}
\alias{diff_water}
\title{Calculate the difference in wawter volume given temperature and humidity inside and outside}
\usage{
diff_water(T_out, h_out, T_in, h_in, volume = 300)
}
\arguments{
\item{T_out}{temperature outside in degress celsius}

\item{h_out}{relative humidity inside in percent}

\item{T_in}{temperature inside in degress celsius}

\item{h_in}{relative humidity outside in percent}

\item{volume}{Volume that is exchanged}
}
\value{
difference in water content before and after changing the air
}
\description{
Calculate the difference in wawter volume given temperature and humidity inside and outside
}
\examples{
diff_water(8, 80, 19, 70)
}
