% Generated by roxygen2 (4.0.2): do not edit by hand
\name{more}
\alias{more}
\title{Display the contents of a file to the R terminal}
\usage{
more(file, n = 500)
}
\arguments{
\item{file}{Text string giving the file name}

\item{n}{Integer specifying the maximum number of lines to display.  This is passed to the
\code{n} argument in \code{\link{readLines}}.}
}
\value{
Invisibly returns the contents of the file are displayed on the R console
}
\description{
Display the contents of a file to the R terminal, like the \code{more} command in Unix
}
\examples{
cat("Here's a file\\n", "with a few lines\\n", "to read.\\n", sep = "", file = "tmpFile.txt")
splashFile("tmpFile.txt")
unlink("tmpFile.txt")
}
\seealso{
\code{\link{readLines}}
}

