% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rma.R
\name{rma}
\alias{rma}
\title{Remove all objects from the global environment}
\usage{
rma()
}
\description{
Remove all objects from the global environment
}
\details{
Alias for \code{rm(list=ls())}.
}
\author{
Landon Sego
}
\keyword{misc}

