% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{Smisc-package}
\alias{Smisc}
\alias{Smisc-package}
\title{Sego Miscellaneous}
\description{
Sego Miscellaneous:  A collection of functions for statistical computing and data manipulation
}
\details{
\tabular{ll}{
Package: \tab Smisc\cr
Type: \tab Package\cr
Version: \tab 0.2.7\cr
Date: \tab 2015-01-31\cr
License: \tab LICENSE\cr
}
}
\examples{
help(package = Smisc)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

