% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smisc-package.R
\docType{package}
\name{Smisc-package}
\alias{Smisc}
\alias{Smisc-package}
\title{Smisc: Sego Miscellaneous}
\description{
A collection of functions for statistical computing and data manipulation
}
\details{
\tabular{ll}{
Package: \tab Smisc\cr
Type: \tab Package\cr
Version: \tab 0.3.3\cr
Date: \tab 2016-05-09\cr
License: \tab file LICENSE\cr
}
}
\examples{
help(package = Smisc)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

