\name{Smisc-package}
\alias{Smisc-package}
\alias{Smisc}
\docType{package}
\title{
Miscellaneous useful functions
}
\description{
Right now, this is just a collection of useful functions for me. It may be more at a later date.
}
\details{
\tabular{ll}{
Package: \tab Smisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-26\cr
License: \tab GPL-2 \cr
}
An overview of how to use the package, including the most important functions
}
\author{
Sean Murphy

Maintainer: Sean Murphy <seanchrismurphy@gmail.com>
}
