% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{Smisc}
\alias{Smisc}
\alias{Smisc-package}
\title{A collection of miscellaneous functions}
\description{
A collection of miscellaneous functions
}
\examples{
help(package = Smisc)
}
\author{
Landon Sego

Maintainer: Landon Sego <Landon.Sego@pnnl.gov>
}
\keyword{package}

