% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sourceDir}
\alias{sourceDir}
\title{'sources' all files with '.R' or '.r' extensions in a directory}
\usage{
sourceDir(directory, recursive = FALSE, tryCatch = TRUE, ...)
}
\arguments{
\item{directory}{Character string indicating the path of the directory
containing the R source files}

\item{recursive}{\code{=TRUE} descends into subdirectories of
\code{directory}}

\item{tryCatch}{if \code{TRUE}, sourcing is protected in a try catch, i.e.,
if there is an error, \code{sourceDir} will continue to the next file.  If
\code{FALSE}, \code{sourceDir} will stop if a file sources with an error.}

\item{\dots}{Additional arguments to \code{source} function}
}
\value{
Invisibly returns a character vector containing the files that were
identified for sourcing.

Also prints a message indicating whether each file was sourced correctly or
not.
}
\description{
'sources' all files with '.R' or '.r' extensions in a directory
}
\examples{
# Source R files in the current directory
sourceDir(".")

# Source R files in the directory "c:\\Work\\myRcode"
# sourceDir("c:/Work/myRcode")
}
\author{
Landon Sego
}
\keyword{misc}

